+++ src/compas_fab/robots/trajectory.py
@@ -6,6 +6,7 @@ from __future__ import print_function
 from compas_fab.robots import AttachedCollisionMesh
 from compas_fab.robots.time_ import Duration
 from compas_fab.robots.configuration import Configuration
+from compas_fab.robots.configuration import FixedLengthList
 
 __all__ = [
     'JointTrajectoryPoint',
@@ -20,15 +21,15 @@ class JointTrajectoryPoint(Configuration):
     A trajectory point is a sub-class of :class:`Configuration` extended
     with acceleration, effort and time from start information.
 
-    Trajectory points are defined either as *values + velocities and
-    accelerations*, or as *values + efforts*.
+    Trajectory points are defined either as *joint_values + velocities and
+    accelerations*, or as *joint_values + efforts*.
 
     Parameters
     ----------
-    values : :obj:`list` of :obj:`float`, optional
+    joint_values : :obj:`list` of :obj:`float`, optional
         Joint values expressed in radians or meters, depending on the respective
         type.
-    types : :obj:`list` of :attr:`compas.robots.Joint.TYPE`, optional
+    joint_types : :obj:`list` of :attr:`compas.robots.Joint.TYPE`, optional
         Joint types, e.g. a :obj:`list` of
         :attr:`compas.robots.Joint.REVOLUTE` for revolute joints.
     velocities : :obj:`list` of :obj:`float`, optional
@@ -42,10 +43,10 @@ class JointTrajectoryPoint(Configuration):
 
     Attributes
     ----------
-    values : :obj:`list` of :obj:`float`
+    joint_values : :obj:`list` of :obj:`float`
         Joint values expressed in radians or meters, depending on the respective
         type.
-    types : :obj:`list` of :attr:`compas.robots.Joint.TYPE`
+    joint_types : :obj:`list` of :attr:`compas.robots.Joint.TYPE`
         Joint types, e.g. a :obj:`list` of
         :attr:`compas.robots.Joint.REVOLUTE` for revolute joints.
     velocities : :obj:`list` of :obj:`float`
@@ -57,24 +58,24 @@ class JointTrajectoryPoint(Configuration):
     time_from_start : :class:`Duration`
         Duration of trajectory point counting from the start.
     positions : :obj:`list` of :obj:`float`
-        Alias of `values`.
+        Alias of `joint_values`.
     data : obj:`dict`
         The data representing the trajectory point.
     """
 
-    def __init__(self, values=None, types=None, velocities=None, accelerations=None, effort=None, time_from_start=None):
-        super(JointTrajectoryPoint, self).__init__(values, types)
-        self.velocities = velocities or len(self.values) * [0.]
-        self.accelerations = accelerations or len(self.values) * [0.]
-        self.effort = effort or len(self.values) * [0.]
+    def __init__(self, joint_values=None, joint_types=None, velocities=None, accelerations=None, effort=None, time_from_start=None, joint_names=None):
+        super(JointTrajectoryPoint, self).__init__(joint_values, joint_types, joint_names)
+        self.velocities = velocities or len(self.joint_values) * [0.]
+        self.accelerations = accelerations or len(self.joint_values) * [0.]
+        self.effort = effort or len(self.joint_values) * [0.]
         self.time_from_start = time_from_start or Duration(0, 0)
 
     def __str__(self):
         """Return a human-readable string representation of the instance."""
         vs = '%.' + self._precision
         return 'JointTrajectoryPoint(({}), {}, ({}), ({}), ({}), {})'.format(
-            ', '.join(vs % i for i in self.values),
-            tuple(self.types),
+            ', '.join(vs % i for i in self.joint_values),
+            tuple(self.joint_types),
             ', '.join(vs % i for i in self.velocities),
             ', '.join(vs % i for i in self.accelerations),
             ', '.join(vs % i for i in self.effort),
@@ -83,8 +84,8 @@ class JointTrajectoryPoint(Configuration):
 
     @property
     def positions(self):
-        """:obj:`list` of :obj:`float` : Alias of `values`."""
-        return self.values
+        """:obj:`list` of :obj:`float` : Alias of `joint_values`."""
+        return self.joint_values
 
     @property
     def velocities(self):
@@ -93,11 +94,11 @@ class JointTrajectoryPoint(Configuration):
 
     @velocities.setter
     def velocities(self, velocities):
-        if len(self.values) != len(velocities):
+        if len(self.joint_values) != len(velocities):
             raise ValueError('Must have {} velocities, but {} given.'.format(
-                len(self.values), len(velocities)))
+                len(self.joint_values), len(velocities)))
 
-        self._velocities = velocities
+        self._velocities = FixedLengthList(velocities)
 
     @property
     def accelerations(self):
@@ -106,11 +107,11 @@ class JointTrajectoryPoint(Configuration):
 
     @accelerations.setter
     def accelerations(self, accelerations):
-        if len(self.values) != len(accelerations):
+        if len(self.joint_values) != len(accelerations):
             raise ValueError('Must have {} accelerations, but {} given.'.format(
-                len(self.values), len(accelerations)))
+                len(self.joint_values), len(accelerations)))
 
-        self._accelerations = accelerations
+        self._accelerations = FixedLengthList(accelerations)
 
     @property
     def effort(self):
@@ -119,11 +120,11 @@ class JointTrajectoryPoint(Configuration):
 
     @effort.setter
     def effort(self, effort):
-        if len(self.values) != len(effort):
+        if len(self.joint_values) != len(effort):
             raise ValueError('Must have {} efforts, but {} given.'.format(
-                len(self.values), len(effort)))
+                len(self.joint_values), len(effort)))
 
-        self._effort = effort
+        self._effort = FixedLengthList(effort)
 
     @property
     def data(self):
@@ -143,11 +144,11 @@ class JointTrajectoryPoint(Configuration):
 
     @data.setter
     def data(self, data):
-        self.values = data.get('values') or []
-        self.types = data.get('types') or []
-        self.velocities = data.get('velocities') or []
-        self.accelerations = data.get('accelerations') or []
-        self.effort = data.get('effort') or []
+        self._joint_values = FixedLengthList(data.get('joint_values') or [])
+        self._joint_types = FixedLengthList(data.get('joint_types') or [])
+        self._velocities = FixedLengthList(data.get('velocities') or [])
+        self._accelerations = FixedLengthList(data.get('accelerations') or [])
+        self._effort = FixedLengthList(data.get('effort') or [])
         self.time_from_start = Duration.from_data(data.get('time_from_start') or {})
 
     def copy(self):
@@ -179,48 +180,6 @@ class JointTrajectoryPoint(Configuration):
         self.check_joint_names()
         return dict(zip(self.joint_names, self.effort))
 
-    def merge(self, other):
-        """Merge the ``JointTrajectoryPoint`` with another ``JointTrajectoryPoint`` in place
-        along joint names.  The other ``JointTrajectoryPoint`` takes precedence over this
-        ``JointTrajectoryPoint`` in case a joint value is present in both.
-
-        Note
-        ----
-            Caution: ``joint_names`` may be rearranged.
-
-        Parameters
-        ----------
-        other : :class:`JointTrajectoryPoint`
-            The ``JointTrajectoryPoint`` to be merged.
-
-        Raises
-        ------
-        :exc:`ValueError`
-            If the ``JointTrajectoryPoint`` or the other ``JointTrajectoryPoint`` does not specify
-            joint names for all joint values.
-        """
-        _joint_dict = self.joint_dict
-        _joint_dict.update(other.joint_dict)
-
-        _type_dict = self.type_dict
-        _type_dict.update(other.type_dict)
-
-        _velocity_dict = self.velocity_dict
-        _velocity_dict.update(other.velocity_dict)
-
-        _acceleration_dict = self.acceleration_dict
-        _acceleration_dict.update(other.acceleration_dict)
-
-        _effort_dict = self.effort_dict
-        _effort_dict.update(other.effort_dict)
-
-        self.joint_names = list(_joint_dict.keys())
-        self.values = [_joint_dict[name] for name in self.joint_names]
-        self.types = [_type_dict[name] for name in self.joint_names]
-        self.velocities = [_velocity_dict[name] for name in self.joint_names]
-        self.accelerations = [_acceleration_dict[name] for name in self.joint_names]
-        self.effort = [_effort_dict[name] for name in self.joint_names]
-
     def merged(self, other):
         """Get a new ``JointTrajectoryPoint`` with this ``JointTrajectoryPoint`` merged
         with another ``JointTrajectoryPoint``.  The other ``JointTrajectoryPoint``
@@ -246,9 +205,29 @@ class JointTrajectoryPoint(Configuration):
             If the ``JointTrajectoryPoint`` or the other ``JointTrajectoryPoint`` does not specify
             joint names for all joint values.
         """
-        jtp = self.copy()
-        jtp.merge(other)
-        return jtp
+        _joint_dict = self.joint_dict
+        _joint_dict.update(other.joint_dict)
+
+        _type_dict = self.type_dict
+        _type_dict.update(other.type_dict)
+
+        _velocity_dict = self.velocity_dict
+        _velocity_dict.update(other.velocity_dict)
+
+        _acceleration_dict = self.acceleration_dict
+        _acceleration_dict.update(other.acceleration_dict)
+
+        _effort_dict = self.effort_dict
+        _effort_dict.update(other.effort_dict)
+
+        joint_names = list(_joint_dict.keys())
+        joint_values = [_joint_dict[name] for name in joint_names]
+        joint_types = [_type_dict[name] for name in joint_names]
+        velocities = [_velocity_dict[name] for name in joint_names]
+        accelerations = [_acceleration_dict[name] for name in joint_names]
+        effort = [_effort_dict[name] for name in joint_names]
+
+        return JointTrajectoryPoint(joint_values, joint_types, velocities, accelerations, effort, joint_names=joint_names)
 
 
 class Trajectory(object):